renamed folders and artifacts from be.ida to be.ida-mediafoudnry to further support distribution via maven central.
