+++ core/src/main/java/be/ida-mediafoundry/jetpack/dictionaryactivation/models/Config.java
@@ -1,0 +1,31 @@
+package be.ida-mediafoundry.jetpack.dictionaryactivation.models;
+
+/**
+ * Datasource configuration object.
+ */
+public class Config {
+
+    private String itemResourceType;
+
+    private Integer offset;
+
+    private Integer limit;
+
+    public Config(String itemResourceType, Integer offset, Integer limit) {
+        this.itemResourceType = itemResourceType;
+        this.offset = offset;
+        this.limit = limit;
+    }
+
+    public String getItemResourceType() {
+        return itemResourceType;
+    }
+
+    public Integer getOffset() {
+        return offset;
+    }
+
+    public Integer getLimit() {
+        return limit;
+    }
+}
