--- core/src/main/java/be/ida/jetpack/dictionaryactivation/repositories/DictionaryRepository.java
@@ -1,20 +1,0 @@
-package be.ida.jetpack.dictionaryactivation.repositories;
-
-import org.apache.sling.api.resource.Resource;
-
-import java.util.Iterator;
-
-/**
- * Repository For accessing the Dictionary on repository level.
- */
-public interface DictionaryRepository {
-
-    /**
-     * Get dictionaries.
-     *
-     * @return Iterator of resources.
-     * @throws Exception
-     */
-    Iterator<Resource> getDictionaries();
-
-}
