+++ pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>be.ida.jetpack</groupId>
+    <groupId>be.ida-mediafoundry.jetpack</groupId>
     <artifactId>dictionaryactivation</artifactId>
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
