+++ ui.apps/pom.xml
@@ -6,7 +6,7 @@
     <!-- P A R E N T  P R O J E C T  D E S C R I P T I O N                      -->
     <!-- ====================================================================== -->
     <parent>
-        <groupId>be.ida.jetpack</groupId>
+        <groupId>be.ida-mediafoundry.jetpack</groupId>
         <artifactId>dictionaryactivation</artifactId>
         <version>1.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
@@ -86,7 +86,7 @@
                     <group>Jetpack</group>
                     <embeddeds>
                         <embedded>
-                            <groupId>be.ida.jetpack</groupId>
+                            <groupId>be.ida-mediafoundry.jetpack</groupId>
                             <artifactId>dictionaryactivation.core</artifactId>
                             <target>/apps/jetpack/dictionary-activation/install</target>
                         </embedded>
@@ -106,7 +106,7 @@
     <!-- ====================================================================== -->
     <dependencies>
         <dependency>
-            <groupId>be.ida.jetpack</groupId>
+            <groupId>be.ida-mediafoundry.jetpack</groupId>
             <artifactId>dictionaryactivation.core</artifactId>
             <version>${project.version}</version>
         </dependency>
