+++ ui.apps/src/main/content/jcr_root/apps/jetpack/dictionary-activation/components/resourceconfig/resourceconfig.html
@@ -1,4 +1,4 @@
-<tr data-sly-use.model="be.ida.jetpack.dictionaryactivation.models.Dictionary"
+<tr data-sly-use.model="be.ida-mediafoundry.jetpack.dictionaryactivation.models.Dictionary"
     class="foundation-collection-item foundation-collection-navigator card-jetpack-dictionary-activation"
     is="coral-table-row"
     data-path="${model.path}"
