+++ ui.apps/src/main/content/jcr_root/apps/jetpack/dictionary-activation/components/resourcesdatasource/resourcesdatasource.jsp
@@ -2,7 +2,7 @@
 %><%@include file="/libs/granite/ui/global.jsp"%><%
 %><%@page session="false"
           import="com.adobe.granite.ui.components.ds.DataSource,
-          be.ida.jetpack.dictionaryactivation.services.DictionaryDataSourceService"%>
+          be.ida-mediafoundry.jetpack.dictionaryactivation.services.DictionaryDataSourceService"%>
 <%
     DictionaryDataSourceService service = (DictionaryDataSourceService)sling.getService(DictionaryDataSourceService.class);
     DataSource dataSource = service.getDataSource(request, cmp, resource);
