+++ ui.apps/src/main/content/jcr_root/apps/jetpack/dictionary-activation/config/org.apache.sling.serviceusermapping.impl.ServiceUserMapperImpl.amended-dictionary-activation.xml
@@ -3,4 +3,4 @@
       xmlns:jcr="http://www.jcp.org/jcr/1.0"
       jcr:primaryType="sling:OsgiConfig"
       user.default="jetpack-dictionary-activation"
-      user.mapping="[be.ida.jetpack.dictionaryactivation.core=jetpack-dictionary-activation]" />
+      user.mapping="[be.ida-mediafoundry.jetpack.dictionaryactivation.core=jetpack-dictionary-activation]" />
